# Performance improvements to TagSerializer