@marcotc 2020-10-26
👋 @remeh @kbogtob, from across the virtual room! I couldn't find who are the canonical maintainers of this repo, so I guessed you two based on recent history :)
I have at least one other PR to contribute, so let me know if there's anything I can do to make the process smoother.